Restructure directories to be more clear about file types:     - Models is for `.ts` files that are not React modules in any sense     - Components is *generally* for `.tsx` React components/nodes
Removed routes, I think the website is small enough
that it is not currently needed
