+++ frontend/src/components/stats/GraphExample.tsx
