+++ frontend/src/components/stats/LineChart.tsx
