+++ frontend/src/models/PromptTypes.ts
