--- frontend/src/routes/StatsRoute.css
@@ -1,3 +1,0 @@
-.grid-auto-cols {
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-}
